Optimizing fitting voxels
When you are looping through voxels tmpOut is the size of *all* voxels and you are using ind to select the correct index for each non-masked voxel. This generates a very large array. A potentially more efficient approach is to just pass a matrix with the *non-masked voxels only* to the function f. And the rebuild the parameter maps based on indices later in the code. For my stand-alone tests…this yields a little less than a two-fold decrease in comp time